requirements.txt
numpy
pandas
psycopg2
python-dotenv
scikit-learn


#-e . # Automatically install the setup.py file 
     # ie. python setup.py install